Refactor Highlighters
We changed the Highlighter class to HighlighterBase.  The logic for
HighlightHereNode has been abstracted and generalized into
HighlighterBase.  HighlightHereParentNode now inherits from
HighlightHereNode.  There is a breaking change, `highlight_here` used to
concatenate all arguments into the `highlight_class`, but now accepts
the class as the first option and the rest of the options are stored in
`self.options`.

We started writing documentation for HighlighterBase, but haven't
finished.  These changes haven't been tested either.

NEW CHANGES:
- Tests for new template tag refactors
- Add ability to pass in a context variable as a path argument

Signed-off-by: Rocky Meza <rockymeza@gmail.com>