Model inheritance now works with QuerySetManager
When subclassing a model, Django copies all managers from the parent
onto the child. QuerySetManager was providing a `__getattr__` method
that proxied _all_ attributes onto a newly created queryset. This causes
an error when trying to `copy.copy` a manager, because querysets provide
a `__getstate__` method. Fixed by not proxying any special methods to
the queryset.

Incidentally, don't proxy `delete` either because you're not supposed be
to able to call `delete` on a manager.