A RedirectFallbackMiddleware that works with Mezzanine
CommonMiddleware and Mezzanine's urlpatterns don't interact well.
Mezzanine has a url for '.*/', so CommonMiddleware tries to redirect
every no-slash request to the slash version. This means that
RedirectFallbackMiddleware wouldn't check for redirects, because it's
not a 404.

We should still check for a redirect if this response is due to
CommonMiddleware.